+++ jokers/mandelajoker.lua
@@ -8,10 +8,10 @@ SMODS.Joker{ --Mandela Joker
     loc_txt = {
         ['name'] = 'Mandela Joker',
         ['text'] = {
-            [1] = 'placeholder; give suggestions for abilities',
+            [1] = 'Remembers the {C:attention}correct{} positions of your jokers',
             [2] = '',
             [3] = '{C:inactive}Art by kommittt',
-            [4] = 'Ability by ...{}'
+            [4] = 'Ability by planetext{}'
         },
         ['unlock'] = {
             [1] = 'Unlocked by default.'
@@ -33,5 +33,46 @@ SMODS.Joker{ --Mandela Joker
     unlocked = true,
     discovered = true,
     atlas = 'CustomJokers',
-    pools = { ["jfdj_jfdj_jokers"] = true }
+    pools = { ["jfdj_jfdj_jokers"] = true },
+    
+    calculate = function(self, card, context)
+        if context.before and context.cardarea == G.jokers  then
+            if #G.jokers.cards > 1 then
+                G.jokers:unhighlight_all()
+                G.E_MANAGER:add_event(Event({
+                    trigger = 'before',
+                    func = function()
+                        G.E_MANAGER:add_event(Event({
+                            func = function()
+                                G.jokers:shuffle('aajk')
+                                play_sound('cardSlide1', 0.85)
+                                return true
+                            end,
+                        }))
+                        delay(0.15)
+                        G.E_MANAGER:add_event(Event({
+                            func = function()
+                                G.jokers:shuffle('aajk')
+                                play_sound('cardSlide1', 1.15)
+                                return true
+                            end
+                        }))
+                        delay(0.15)
+                        G.E_MANAGER:add_event(Event({
+                            func = function()
+                                G.jokers:shuffle('aajk')
+                                play_sound('cardSlide1', 1)
+                                return true
+                            end
+                        }))
+                        delay(0.5)
+                        return true
+                    end
+                }))
+            end
+            return {
+                message = "Shuffle!"
+            }
+        end
+    end
 }