+++ changelog.md
@@ -5,5 +5,7 @@
 - Added & implemented John Stencil
 - Added Jokey Forgey
 - Added & implemented Mannequin Joker
+- Implemented Eugene
+- Implemented Mandela Joker
 ## 1.0.0 (23/5/2026)
 - Initial version